Resolved a bug where app caches processed parts of the images too

+++ android/app/src/main/java/com/senior_project/group_1/mobilesr/views/BitmapHelpers.java
@@ -264,6 +264,9 @@ public class BitmapHelpers {
             Log.i("BitmapHelpers", "Saving image on "+savingFolder+" / "+filename);
 
             File file = new File(savingFolder, filename + ".png");
+            if( file.exists() )
+                return GenericFileProvider.getUriForFile(context, context.getApplicationContext().getPackageName() + "", file);
+
             file.createNewFile();
 
             FileOutputStream stream = new FileOutputStream(file);
